fix bugs maps

+++ cube_three_d.h
@@ -6,7 +6,7 @@
 /*   By: scolen <scolen@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/15 15:13:51 by scolen            #+#    #+#             */
-/*   Updated: 2021/02/06 22:40:25 by scolen           ###   ########.fr       */
+/*   Updated: 2021/02/06 22:59:14 by scolen           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,8 +24,8 @@
 
 typedef	struct			s_value_from_map
 {
-	int					resolution_x;
-	int					resolution_y;
+	long long			resolution_x;
+	long long			resolution_y;
 
 	char				*north_texture;
 	char				*south_texture;
@@ -154,7 +154,7 @@ typedef	struct			s_object_on_scene
 	int					color_pixel_sprite;
 }						t_object_on_scene;
 
-int						len_number(int number);
+int						len_number(long long number);
 int						ft_strcmp(const char *str1, const char *str2);
 int						pass_space(char *substr);
 int						count_comma(char *substr);
